requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
requests==2.32.5
beautifulsoup4==4.14.3
pydantic==2.12.5
python-docx==1.2.0
playwright==1.57.0
python-multipart==0.0.9
pycryptodome==3.23.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.46
httpx==0.25.2
openai==1.12.0
python-dotenv==1.0.0
pdfplumber==0.10.3
pdf2image==1.17.0
openpyxl==3.1.2
pandas==2.2.0
